User commit

+++ src/main/java/com/example/demo/repository/ValidationRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entity.ValidationEntity;
 
 @Repository
-public interface ValidationRepo extends JpaRepository<ValidationEntity,int>{
+public interface ValidationRepo extends JpaRepository<ValidationEntity,Long>{
 
 }
